clisqlshell: drop some dependencies on `sql/parser`
This moves `LastLexicalToken` to the more lightweight `scanner`
package.

Release note: None
