+++ pkg/cli/clisqlshell/BUILD.bazel
@@ -23,6 +23,7 @@ go_library(
         "//pkg/sql/parser",
         "//pkg/sql/pgwire/pgcode",
         "//pkg/sql/pgwire/pgerror",
+        "//pkg/sql/scanner",
         "//pkg/sql/sem/builtins",
         "//pkg/util/envutil",
         "@com_github_cockroachdb_errors//:errors",
@@ -48,7 +49,7 @@ go_test(
         "//pkg/cli/clisqlexec",
         "//pkg/security",
         "//pkg/server",
-        "//pkg/sql/parser",
+        "//pkg/sql/scanner",
         "//pkg/testutils/serverutils",
         "//pkg/util/leaktest",
         "//pkg/util/log",
