+++ pkg/sql/parser/scanner_test.go
@@ -420,75 +420,3 @@ func TestSplitFirstStatement(t *testing.T) {
 		})
 	}
 }
-
-func TestLastLexicalToken(t *testing.T) {
-	defer leaktest.AfterTest(t)()
-	tests := []struct {
-		s   string
-		res int
-	}{
-		{
-			s:   "",
-			res: 0,
-		},
-		{
-			s:   " /* comment */ ",
-			res: 0,
-		},
-		{
-			s:   "SELECT",
-			res: SELECT,
-		},
-		{
-			s:   "SELECT 1",
-			res: ICONST,
-		},
-		{
-			s:   "SELECT 1;",
-			res: ';',
-		},
-		{
-			s:   "SELECT 1; /* comment */",
-			res: ';',
-		},
-		{
-			s: `SELECT 1;
-			    -- comment`,
-			res: ';',
-		},
-		{
-			s: `
-				--SELECT 1, 2, 3;
-				SELECT 4, 5
-				--blah`,
-			res: ICONST,
-		},
-		{
-			s: `
-				--SELECT 1, 2, 3;
-				SELECT 4, 5;
-				--blah`,
-			res: ';',
-		},
-		{
-			s:   `SELECT 'unfinished`,
-			res: ERROR,
-		},
-		{
-			s:   `SELECT e'\xaa';`, // invalid token but last token is semicolon
-			res: ';',
-		},
-	}
-
-	for i, tc := range tests {
-		t.Run(fmt.Sprintf("%d", i), func(t *testing.T) {
-			tok, ok := LastLexicalToken(tc.s)
-			if !ok && tok != 0 {
-				t.Fatalf("!ok but nonzero tok")
-			}
-			if tc.res != tok {
-				t.Errorf("expected %d but got %d", tc.res, tok)
-			}
-		})
-	}
-}
