+++ pkg/sql/scanner/scan_test.go
@@ -10,7 +10,12 @@
 
 package scanner
 
-import "testing"
+import (
+	"fmt"
+	"testing"
+
+	"github.com/cockroachdb/cockroach/pkg/sql/lexbase"
+)
 
 func TestHasMultipleStatements(t *testing.T) {
 	testCases := []struct {
@@ -35,5 +40,76 @@ func TestHasMultipleStatements(t *testing.T) {
 		if actual != tc.expected {
 			t.Errorf("%q: expected %v, got %v", tc.in, tc.expected, actual)
 		}
+	}
+}
+
+func TestLastLexicalToken(t *testing.T) {
+	tests := []struct {
+		s   string
+		res int
+	}{
+		{
+			s:   "",
+			res: 0,
+		},
+		{
+			s:   " /* comment */ ",
+			res: 0,
+		},
+		{
+			s:   "SELECT",
+			res: lexbase.SELECT,
+		},
+		{
+			s:   "SELECT 1",
+			res: lexbase.ICONST,
+		},
+		{
+			s:   "SELECT 1;",
+			res: ';',
+		},
+		{
+			s:   "SELECT 1; /* comment */",
+			res: ';',
+		},
+		{
+			s: `SELECT 1;
+			    -- comment`,
+			res: ';',
+		},
+		{
+			s: `
+				--SELECT 1, 2, 3;
+				SELECT 4, 5
+				--blah`,
+			res: lexbase.ICONST,
+		},
+		{
+			s: `
+				--SELECT 1, 2, 3;
+				SELECT 4, 5;
+				--blah`,
+			res: ';',
+		},
+		{
+			s:   `SELECT 'unfinished`,
+			res: lexbase.ERROR,
+		},
+		{
+			s:   `SELECT e'\xaa';`, // invalid token but last token is semicolon
+			res: ';',
+		},
+	}
+
+	for i, tc := range tests {
+		t.Run(fmt.Sprintf("%d", i), func(t *testing.T) {
+			tok, ok := LastLexicalToken(tc.s)
+			if !ok && tok != 0 {
+				t.Fatalf("!ok but nonzero tok")
+			}
+			if tc.res != tok {
+				t.Errorf("expected %d but got %d", tc.res, tok)
+			}
+		})
 	}
 }
